requirements.txt
fastapi>=0.100.0
pydantic>=2.0.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
email-validator>=2.1.0
